Lift the closure call-site vocabulary out of the census rule
`CollectionOperation`, `ForwardingCall` and the enclosing-declaration walk
were private to `PureClosureCandidateVisitor`. The inventory rule that
follows has to run the *identical* gate — the same higher-order call sites,
the same "hides a law worth stating" filter, the same forwarding check — or
the two numbers stop being complements and neither can be read against the
other.

A pure move plus visibility; no logic changes. This project has found the same
shape four times over in the architecture rules: one rule privately holding
the vocabulary its twin needs, and reporting something the twin already knew
was not a finding. Doing it before the second consumer exists is the cheap
version of that fix.
